Replace EOL PyPy 3.9/3.10 with PyPy 3.11 in CI matrix
